chahua: 重构 renderer.js（三）—— 抽出 room_list 模块
侧栏房间列表 + P9「进行中」后台房徽标抽到 room_list.js。「进行中」后台房集合
（backgroundActiveRooms）全封进模块，与 message_filter 的 filter 状态同口径 ——
renderer 只经 render（room_info 来时按权威快照重建 + 渲染）/ onBackgroundMilestone
（后台里程碑事件增量维护）两个出口交互，handleBackgroundMilestone 整个移入。

renderer.js 1422 → 1331 行（自原始 1671 累计 −340）。行为零变化，node --check 通过。
